refactor(zetaclient): share outbound amount guard
