+++ zetaclient/chains/base/amount_test.go
@@ -1,4 +1,4 @@
-package signer
+package base
 
 import (
 	"math/big"
@@ -13,7 +13,7 @@ import (
 func TestOutboundAmountUint64(t *testing.T) {
 	t.Run("returns amount within uint64 range", func(t *testing.T) {
 		maxUint64 := ^uint64(0)
-		amount, err := outboundAmountUint64(sdkmath.NewUint(maxUint64))
+		amount, err := OutboundAmountUint64(sdkmath.NewUint(maxUint64))
 		require.NoError(t, err)
 		require.Equal(t, maxUint64, amount)
 	})
@@ -21,7 +21,7 @@ func TestOutboundAmountUint64(t *testing.T) {
 	t.Run("fails if amount exceeds uint64 range", func(t *testing.T) {
 		amount := sdkmath.NewUintFromBigInt(new(big.Int).Lsh(big.NewInt(1), 64))
 
-		_, err := outboundAmountUint64(amount)
+		_, err := OutboundAmountUint64(amount)
 		require.ErrorIs(t, err, cctypes.ErrInvalidWithdrawalAmount)
 	})
 }
