+++ zetaclient/chains/solana/signer/execute.go
@@ -17,6 +17,7 @@ import (
 	"github.com/zeta-chain/node/pkg/chains"
 	contracts "github.com/zeta-chain/node/pkg/contracts/solana"
 	"github.com/zeta-chain/node/x/crosschain/types"
+	"github.com/zeta-chain/node/zetaclient/chains/base"
 	zctx "github.com/zeta-chain/node/zetaclient/context"
 )
 
@@ -161,7 +162,7 @@ func (signer *Signer) createMsgExecute(
 	// zero out the amount if cancelTx is set. It's legal to withdraw 0 lamports through the gateway.
 	if !cancelTx {
 		var err error
-		amount, err = outboundAmountUint64(params.Amount)
+		amount, err = base.OutboundAmountUint64(params.Amount)
 		if err != nil {
 			return nil, nil, err
 		}
