+++ zetaclient/chains/solana/signer/withdraw.go
@@ -11,6 +11,7 @@ import (
 	"github.com/zeta-chain/node/pkg/chains"
 	contracts "github.com/zeta-chain/node/pkg/contracts/solana"
 	"github.com/zeta-chain/node/x/crosschain/types"
+	"github.com/zeta-chain/node/zetaclient/chains/base"
 )
 
 // prepareWithdrawTx prepares withdraw outbound
@@ -58,7 +59,7 @@ func (signer *Signer) createMsgWithdraw(
 	// zero out the amount if cancelTx is set. It's legal to withdraw 0 lamports through the gateway.
 	if !cancelTx {
 		var err error
-		amount, err = outboundAmountUint64(params.Amount)
+		amount, err = base.OutboundAmountUint64(params.Amount)
 		if err != nil {
 			return nil, nil, err
 		}
